+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Deprecated
 ### Removed
 - JsonKraken 1.0 Source
+- kSock (`klib.net.socket.kSock`)
 - Directory (`klib.files.Directory`)
 ### Fixed
 ### Security