+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - JsonKraken 2.0 dependency
 ### Changed
+- Upgraded Kotlin from 1.3.61 to 1.3.70
 ### Deprecated
 ### Removed
 - JsonKraken 1.0 Source