+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Status number from `klib.files.Directory2.DirectoryError`
 ### Fixed
 - FFDB now checks if the File exists before writing/reading
+- FFDB now checks if the FIle is empty before reading
 ### Security
 
 ## 4.1.0 - 25.02.2020 - Release 24